// frontend/src/client/types.gen.ts
export type GroupRead = {
    /**
     * Name
     */
    name: string;
    /**
     * Course
     */
    course: number;
    /**
     * Institute
     */
    institute: string;
    /**
     * Id
     */
    id: number;
    /**
     * Createdat
     */
    createdAt: string;
    /**
     * Updatedat
     */
    updatedAt: string;
};
// ...
export type LessonById = {
    /**
     * Id
     */
    id: number;
    /**
     * Timeid
     */
    timeId: number;
    /**
     * Dayofweek
     */
    dayOfWeek: number;
    /**
     * Type
     */
    type: string;
    group: GroupRead;
    subject: SubjectRead;
    teacher: TeacherRead;
    room: RoomReadMinimal;
};
// ...
export type RoomReadMinimal = {
    /**
     * Name
     */
    name: string;
    /**
     * Floor
     */
    floor: number;
    /**
     * Capacity
     */
    capacity: number;
    /**
     * Status
     *
     * 0 - Not available, 1 - Available
     */
    status: number;
    /**
     * Buildingid
     */
    buildingId: number;
    /**
     * Id
     */
    id: number;
};
// ...
export type SubjectRead = {
    /**
     * Name
     */
    name: string;
    /**
     * Semester
     */
    semester: number;
    /**
     * Totalhours
     */
    totalHours: number;
    /**
     * Isoptional
     */
    isOptional: boolean;
    /**
     * Id
     */
    id: number;
    /**
     * Createdat
     */
    createdAt: string;
    /**
     * Updatedat
     */
    updatedAt: string;
};
// ...
export type TeacherRead = {
    /**
     * Firstname
     */
    firstName: string;
    /**
     * Middlename
     */
    middleName: string | null;
    /**
     * Lastname
     */
    lastName: string;
    /**
     * Email
     */
    email: string | null;
    /**
     * Phone
     */
    phone: string | null;
    /**
     * Department
     */
    department: string;
    /**
     * Title
     */
    title: string;
    /**
     * Id
     */
    id: number;
    /**
     * Fullname
     */
    fullName: string;
    /**
     * Createdat
     */
    createdAt: string;
    /**
     * Updatedat
     */
    updatedAt: string;
};